Terrain: more double to float

+++ OpenSim/Region/CoreModules/World/Terrain/Modifiers/FillModifier.cs
@@ -80,10 +80,10 @@ namespace OpenSim.Region.CoreModules.World.Terrain.Modifiers
             return val;
         }
 
-        public override double operate(double[,] map, TerrainModifierData data, int x, int y)
+        public override float operate(float[,] map, TerrainModifierData data, int x, int y)
         {
-            double factor = this.computeBevel(data, x, y);
-            double result = data.elevation - (data.elevation - data.bevelevation) * factor;
+            float factor = this.computeBevel(data, x, y);
+            float result = data.elevation - (data.elevation - data.bevelevation) * factor;
             return result;
         }
